refactor: add cases package shims for case workflows
