+++ dfode_kit/__init__.py
@@ -30,8 +30,8 @@ _ATTRIBUTE_MODULES = {
     "inverse_BCT": ("dfode_kit.utils", "inverse_BCT"),
     "BCT_torch": ("dfode_kit.utils", "BCT_torch"),
     "inverse_BCT_torch": ("dfode_kit.utils", "inverse_BCT_torch"),
-    "gather_species_arrays": ("dfode_kit.df_interface.sample_case", "gather_species_arrays"),
-    "df_to_h5": ("dfode_kit.df_interface.sample_case", "df_to_h5"),
+    "gather_species_arrays": ("dfode_kit.cases.sampling", "gather_species_arrays"),
+    "df_to_h5": ("dfode_kit.cases.sampling", "df_to_h5"),
     "touch_h5": ("dfode_kit.data_operations.h5_kit", "touch_h5"),
     "get_TPY_from_h5": ("dfode_kit.data_operations.h5_kit", "get_TPY_from_h5"),
     "advance_reactor": ("dfode_kit.data_operations.h5_kit", "advance_reactor"),
